+++ react/__mocks__/vtex.native-types/index.tsx
@@ -1,9 +1,1 @@
-import { InjectedIntl } from 'react-intl'
-
-export const formatIOMessage = ({
-  id,
-  intl,
-}: {
-  id: string
-  intl: InjectedIntl
-}) => intl.formatMessage({ id })
+export const formatIOMessage = ({ id }: { id: string }) => id