+++ react/__test__/index.test.tsx
@@ -199,6 +199,7 @@ teste|abc
 
     expect(asFragment()).toMatchSnapshot()
   })
+
   it('should render with two different heading levels', () => {
     const component = render(
       <RichText
@@ -212,6 +213,7 @@ teste|abc
     expect(component).toBeDefined()
     expect(component.asFragment()).toMatchSnapshot()
   })
+
   it('should render list', () => {
     const component = render(
       <RichText
@@ -224,6 +226,28 @@ teste|abc
     )
     expect(component).toBeDefined()
     expect(component.asFragment()).toMatchSnapshot()
+  })
+
+  it('should sanitize the font prop', () => {
+    const typography = 't-heading-1'
+    const { container } = render(<RichText {...defaultProps} text="foo" font={`${typography} foo`} />)
+
+    const element = container.querySelector(`.${typography}`)
+    const notFound = container.querySelector('.foo')
+
+    expect(element).toBeTruthy()
+    expect(notFound).toBeFalsy()
+  })
+
+  it('should sanitize the textColor prop', () => {
+    const color = 'c-muted-1'
+    const { container } = render(<RichText {...defaultProps} text="foo" textColor={`${color} foo`} />)
+
+    const element = container.querySelector(`.${color}`)
+    const notFound = container.querySelector('.foo')
+
+    expect(element).toBeTruthy()
+    expect(notFound).toBeFalsy()
   })
 })