+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Avoid passing classes that are not typography tokens to the `font` prop.
+- Avoid passing classes that are not color tokens to the `textColor` prop.
+
 ## [0.7.1] - 2019-06-07
 ### Changed
 - Make headers convert to html `<hx>` classes and not span.